+++ CHANGELOG.md
@@ -9,10 +9,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Added ARIA import
 - `quadtree` added GeoJSON export
+- spool: covariance progress bar
 
 ### Fixed
 - `spool` added absolute degree to cursor position
-- improved LiCSAR import
+- bugfixes LiCSAR import
 
 ## [1.2.3] 17. July 2019